refactor(frontend)!: separate HDL frontend modules
