+++ crates/celox-frontend/src/symbolic/bitaccess.rs
@@ -8,7 +8,7 @@ use veryl_analyzer::ir::{
 use veryl_analyzer::value::Value;
 use veryl_parser::token_range::TokenRange;
 
-use crate::{ParserError, types::extend_resolved_dims};
+use crate::{ParserError, symbolic::types::extend_resolved_dims};
 
 /// Extract a compile-time constant value for Celox 4-state encoding.
 ///
