+++ crates/celox-frontend/src/symbolic/case.rs
