+++ crates/celox-frontend/src/symbolic/context_width.rs
