+++ crates/celox-frontend/src/symbolic/global_ff.rs
