+++ crates/celox-frontend/src/symbolic/logic_tree/effect.rs
