+++ crates/celox-frontend/src/symbolic/logic_tree/state.rs
