+++ crates/celox-frontend/src/symbolic/registry.rs
