+++ crates/celox-frontend/src/veryl/loop_provenance.rs
