+++ crates/celox/src/parser/module.rs
@@ -1,1 +1,1 @@
-pub use celox_frontend::module::*;
+pub use celox_frontend::veryl::module::*;
